GWL-51 GWL-52 GWL-53 GWL-54 done

+++ app/Http/Controllers/API/PurchaseRequestAPIController.php
@@ -958,9 +958,10 @@ class PurchaseRequestAPIController extends AppBaseController
         }
 
         $documentApproval = DocumentApproved::where('companySystemID', $purchaseRequest->companySystemID)
-            ->where('documentSystemCode', $purchaseRequest->purchaseRequestID)
-            ->where('documentSystemID', $purchaseRequest->documentSystemID)
-            ->get();
+                                                ->where('documentSystemCode', $purchaseRequest->purchaseRequestID)
+                                                ->where('documentSystemID', $purchaseRequest->documentSystemID)
+                                                ->where('approvedYN', -1)
+                                                ->get();
 
         foreach ($documentApproval as $da) {
             $emails[] = array('empSystemID' => $da->employeeSystemID,
@@ -972,6 +973,9 @@ class PurchaseRequestAPIController extends AppBaseController
         }
 
         $sendEmail = \Email::sendEmail($emails);
+        if (!$sendEmail["success"]) {
+            return $this->sendError($sendEmail["message"],500);
+        }
 
         return $this->sendResponse($purchaseRequest, 'Purchase Request successfully canceled');
 
@@ -1038,6 +1042,7 @@ class PurchaseRequestAPIController extends AppBaseController
         $documentApproval = DocumentApproved::where('companySystemID', $purchaseRequest->companySystemID)
                                             ->where('documentSystemCode', $purchaseRequest->purchaseRequestID)
                                             ->where('documentSystemID', $purchaseRequest->documentSystemID)
+                                            ->where('approvedYN', -1)
                                             ->get();
 
         foreach ($documentApproval as $da) {
@@ -1052,6 +1057,10 @@ class PurchaseRequestAPIController extends AppBaseController
         }
 
         $sendEmail = \Email::sendEmail($emails);
+        if (!$sendEmail["success"]) {
+            return $this->sendError($sendEmail["message"],500);
+        }
+
         DocumentApproved::destroy($ids_to_delete);
 
         return $this->sendResponse($purchaseRequest, 'Purchase Request successfully return back to amend');
